\name{in2extRemes-package}
\alias{in2extRemes-package}
\alias{in2extRemes}
\docType{package}
\title{
Graphical User Interface Dialog Window for EVA
}
\description{
Open the main graphical user interface dialog window for extreme value analysis (EVA).
}
\usage{
in2extRemes()
}
\details{

This package provides point-and-click windows for performing extreme value analysis using the package \pkg{extRemes} version >= 2.0.

The function \code{in2extRemes} is used to open the main dialog window from the command line.

}
\value{

No value is returned.  A GUI window is opened from which other functions can be called that might return various object types.

}
\references{
\url{http://www.assessment.ucar.edu/toolkit}
}
\author{
Eric Gilleland
}

\seealso{
\code{\link{extRemes}}
}
\examples{
\dontrun{
in2extRemes()
}
}
\keyword{ package }
\keyword{ utilities }
